pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
ffmpeg-python = "^0.2.0"
numpy = "^2.0.0"
opencv-python = "^4.8.1"
tqdm = "^4.66.1"

[tool.poetry.dev-dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
